Move logic to call the selectedRole endpoint to the EpsCard
